ajout and amelioration du chat, type user et autres

+++ routes/web.php
@@ -22,22 +22,29 @@ Route::get('/', function () {
 
 Auth::routes(['verify'=>true]);
 
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('dashboard')->middleware('verified');
-Route::get('/files', [App\Http\Controllers\FileController::class, 'index'])->name('admin.files')->middleware('verified');
-Route::get('messages/{id?}', [App\Http\Controllers\MessageController::class,'index'])->name('admin.messages')->middleware('verified');
-Route::get('message/{id}', [App\Http\Controllers\MessageController::class,'show'])->name('admin.message')->middleware('verified');
-Route::get('chat/{id?}',[MessageController::class,'getMessage'])->middleware('verified')->name('admin.message');
+#news route configarition
 
-Route::get('list-patients',[HomeController::class,'listPatients'])->name('admin.list-patients');
-Route::get('list-patient/{id}',[HomeController::class,'getByIdPatient'])->name('admin.list-patient');
-// Route::get('/user/profile/{id}',[HomeController::class,'show'])->name('profile');
-// Route::match(['put', 'patch'], '/user/profile_update/{id}',[UserController::class,'update'])->name('profile_update');
 
-// generete all routes of UserController?
-Route::get('/users/', [UserController::class,'index']);
-Route::get('/users/profile/{id}', [UserController::class,'show'])->name('users.profile')->middleware('verified');
-Route::post('/users', [UserController::class,'store']);
-Route::put('/users/profile/{id}',[UserController::class,'update'])->name('users.update')->middleware('verified');
-Route::delete('/users/{id}', [UserController::class,'destroy']);
-Route::post('users/update',[UserController::class,'updateProfile'])->name('update');
-Route::post('envoi',[MessageController::class,'store'])->name('envoi');
+Route::get('messages',[MessageController::class,'index'])->name('admin.messages')->middleware(['verified','auth']);
+Route::get('conversations',[MessageController::class,'index'])->name('admin.conversations')->middleware(['verified','auth']);
+Route::get('conversations/{user}',[MessageController::class,'show'])->name('admin.conversations.show')->middleware(['verified','auth']);
+Route::post('conversations/{user}',[MessageController::class,'store'])->middleware(['verified','auth']);
+Route::delete('conversations/{id}',[MessageController::class,'destroy'])->name('delete')->middleware(['verified','auth']);
+
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('dashboard')->middleware(['verified','auth']); 
+Route::get('list-patients',[HomeController::class,'listPatients'])->name('admin.list-patients')->middleware(['verified','auth']);
+Route::get('list-patient/{all}',[HomeController::class,'getByIdPatient'])->name('admin.list-patient')->middleware(['verified','auth']);
+Route::delete('list-patient/{user}',[HomeController::class,'destroy'])->name('deleted')->middleware(['verified','auth']);
+
+
+Route::get('/files', [App\Http\Controllers\FileController::class, 'index'])->name('admin.files')->middleware(['verified','auth']); 
+
+
+Route::get('/users/', [UserController::class,'index'])->middleware(['verified','auth']);
+Route::get('/users/profile/{id}', [UserController::class,'show'])->name('users.profile')->middleware(['verified','auth']);
+Route::post('/users', [UserController::class,'store'])->middleware(['verified','auth']);
+Route::put('/users/profile/{id}',[UserController::class,'update'])->name('users.update')->middleware(['verified','auth']);
+Route::delete('/users/{id}', [UserController::class,'destroy'])->middleware(['auth','verified']);
+Route::post('users/update',[UserController::class,'updateProfile'])->name('update')->middleware(['verified','auth']);
+Route::post('envoi',[MessageController::class,'store'])->name('envoi')->middleware(['verified','auth']);
